Provide flat iteration space in TensorView operations
It might be easier to iterate over a flat space than
to iterate over the NDIM space. This renames foreach_idx
to foreach_idxs and introduces foreach_idx that passes
a single index in the range of `[0...size)`. Also adds
`operator[]()` to be used with the flat index space.

I'm not sure how well this work for `SliceTensorViews`.
We might want to add specialized overloads for slice
assignment both ways.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>