Properly propagate leaf information
Carefully adjust leaf and child leaf information to make sure that
nodes always carry the correct information on whether they are
leafs and all their children are leafs. This is especially tryue for
empty nodes, which may have to combine child node information
from multiple input nodes in the case of gaxpy.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>