Optimize GEMM for larger threadblocks
In 3D, the mTxm in transform performs a GEMM that where A and C are
dimj^2 x dimj and B is dimjxdimj (with dimj = dimk). We can thus
allow for a third dimension in the threadblock to operate on multiple
dimjxdimj tiles concurrently. This enables the use of more threads.
The optimized version also removes two loops since we assume that the
x and y dimension of the threadblock are the same as dimj and dimk.


Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>