Increase number of threads per block
We can run with up to KxKxmin(K,1024/K**2) threads in each block.
This is straigh-forward for assignments (via the flat iteration space)
and for reductions.
For the tall-and-skinny (100x10 or (400x20) matrices in transform
we can distribute the i-dimension across both the y and z dimension
of the thread-block.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>