Block mTxm on 32 threads in X
Block A matrix but don't transpose it in shared memory.
That way (and with 32 threads in X) we can utilize shared memory broadcast.
We may have to revisit this though because some threads are now idle.
Maybe a better solution is to rearrange the elements in block_a so that
successive threads can access their elements sequentially but that will
require more shared memory / more blocks.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>